+++ pkg/parser/parser_type.go
@@ -1,0 +1,35 @@
+package parser
+
+// parseType parses `num[]{}` into `MAP ARRAY NUM` inverting the order.
+func (p *Parser) parseType() *Type {
+	result := p.parseBasicType()
+	if result == ILLEGAL_TYPE {
+		return result
+	}
+	return p.parseSubType(result)
+}
+
+func (p *Parser) parseBasicType() *Type {
+	tt := p.cur.TokenType()
+	t := basicTypeName(tt)
+	p.advance()
+	if t == ILLEGAL {
+		return ILLEGAL_TYPE
+	}
+	return &Type{Name: t}
+}
+
+func (p *Parser) parseSubType(parent *Type) *Type {
+	tt := p.cur.TokenType()
+	typeName := compositeTypeName(tt)
+	if typeName == ILLEGAL { // we have moved passed the type declaration
+		return parent
+	}
+	if !matchParen(tt, p.peek.Type) {
+		return ILLEGAL_TYPE
+	}
+	p.advance() // advance past opening token `[` or `{`
+	p.advance() // advance past closing token `]` or `}`
+	node := &Type{Name: typeName, Sub: parent}
+	return p.parseSubType(node)
+}
